Implement run_commands/2 from erlangs quickcheck
This allows to create a system map and pass it to run_commands. Until
now i always created a system as extra command in my projects
as the first command. After the run i fetched the system from the state
and closed it.
Now one can simply create the system in the property, bind it with let
and close it with (try ... (finally)) block.